Move the handler tests to jsonrpc module
